feat(trusty-review): opt-in test-coverage grading/gating dimension (closes #1014)
Adds LCOV-based coverage ingestion and a configurable policy that can
lower the letter grade and floor the verdict to REQUEST_CHANGES when
new-code coverage or net coverage drops fall below configurable thresholds.

OFF by default (TRUSTY_REVIEW_COVERAGE_ENABLED is unset → false).
Existing behaviour is 100% unchanged when the policy is disabled.

New config knob:
  TRUSTY_REVIEW_COVERAGE_ENABLED  — enable/disable (default false)
  TRUSTY_REVIEW_MIN_NEW_CODE_PCT  — min new-code coverage pct (default 80)
  TRUSTY_REVIEW_MAX_NET_DROP_PCT  — max allowed net drop in pct-pts (default 1)
  TRUSTY_REVIEW_LCOV_PATH         — path to LCOV report file

Also configurable via [coverage] table in the TOML config file.

Grade/verdict mapping:
  - new-code coverage < min_new_code_pct → floor to REQUEST_CHANGES, grade ≤ D+
  - net coverage drop > max_net_drop_pct → floor to REQUEST_CHANGES, grade ≤ D+
  - BLOCK verdict is never weakened; floor only tightens APPROVE/APPROVE*
  - LLM is informed of coverage data in user message but told not to use it
    for verdict escalation (runner applies the floor deterministically)

Files split to stay under the 500-line cap (#610):
  coverage/{mod,lcov,policy}.rs — LCOV parser, policy, floor application
  pipeline/runner_coverage.rs   — LCOV loading + diff added-line extraction
  pipeline/runner_helpers.rs    — grade/floor derivation, GH meta fetch, abort/finalise
  pipeline/prompt_templates.rs  — system-prompt string constants
  pipeline/prompt_user_msg.rs   — user-message builder
