+++ crates/trusty-review/src/pipeline/prompt.rs
@@ -22,6 +22,7 @@
 //! `prompt_includes_context_blocks`.
 
 use crate::{
+    coverage::CoverageVerdictContrib,
     integrations::{
         analyze_client::{ComplexityHotspot, Smell},
         apex_context::ApexContextResult,
@@ -31,6 +32,12 @@ use crate::{
     models::ReviewResult,
     voice::VoiceConfig,
 };
+
+// System prompt templates are in a separate file to keep this module under the
+// 500-line cap (#610) — the two large prompt constants are ~160 lines combined.
+use super::prompt_templates::{SYSTEM_PROMPT_COVERAGE_GATING, SYSTEM_PROMPT_STOCK};
+// User-message builder extracted to keep this module under the 500-line cap (#610).
+use super::prompt_user_msg::build_user_message;
 
 // ─── Prompt constants ─────────────────────────────────────────────────────────
 
@@ -130,6 +137,14 @@ pub struct ReviewContext {
     /// APEX is disabled (`apex_index` not configured) or no matching docs are
     /// found.  Fail-open: a search error produces an empty vec, never an error.
     pub apex_results: Vec<ApexContextResult>,
+    /// Coverage verdict contribution from the coverage policy (#1014).
+    ///
+    /// `None` when coverage gating is disabled (the default) or when no LCOV
+    /// file was available.  When `Some`, the summary string is injected into the
+    /// user message as an informational block so the LLM can reference it in
+    /// findings; the `floor` is applied deterministically by the runner AFTER
+    /// the LLM response, not by the model itself.
+    pub coverage_contrib: Option<CoverageVerdictContrib>,
 }
 
 // ─── System prompt ────────────────────────────────────────────────────────────
@@ -141,81 +156,34 @@ pub struct ReviewContext {
 /// REQUEST_CHANGES/BLOCK.  Kept as a function for backward compatibility and
 /// for tests that need only the stock text.  For the full 3-layer prompt
 /// (stock → principles → voice) use `build_system_prompt(voice_config)`.
+/// The `coverage_gating_enabled` parameter controls whether the prompt tells
+/// the model that coverage can gate the verdict (#1014).  When `false`, the
+/// stock advisory text ("do not block on coverage") is preserved unchanged.
 /// What: returns a static string; the output-format section uses structured
 /// output language — the provider forces JSON via `response_schema` so the
 /// model need not emit a fenced block.
-/// Test: `system_prompt_contains_policy`.
+/// Test: `system_prompt_contains_policy`, `system_prompt_coverage_gating_on`,
+/// `system_prompt_coverage_gating_off`.
 pub fn reviewer_system_prompt() -> &'static str {
-    r#"You are a senior software engineer performing a pull-request code review.
+    reviewer_system_prompt_with_coverage(false)
+}
 
-## Letter grade (MANDATORY — assign exactly one)
-
-Assign a letter grade on the 13-step scale: A+, A, A-, B+, B, B-, C+, C, C-, D+, D, D-, F.
-
-| Grade band        | Quality signal                                              |
-|-------------------|-------------------------------------------------------------|
-| A+, A, A-         | Excellent to exceptional — clean, correct, well-structured. |
-| B+, B, B-         | Good to solid — acceptable, minor nits only.                |
-| C+, C, C-         | Marginal — notable issues or advisory concerns.             |
-| D+, D, D-         | Poor — significant problems requiring changes before merge. |
-| F                 | Failing — compile error, data corruption, security bypass.  |
-
-Provide a one-line justification in `grade_justification`.
-
-## Verdict (MANDATORY — pick exactly one)
-
-| Verdict         | Grade band      | When to use |
-|-----------------|-----------------|-------------|
-| BLOCK           | F               | Compile error introduced by this diff, data corruption, security/auth bypass. |
-| REQUEST_CHANGES | D+, D, D-       | Confirmed correctness bug, silent data loss, missing required migration/backfill, resource leak, unhandled exception path with real failure consequence. |
-| APPROVE*        | C+, C, C-       | Advisory concern the author may reasonably disagree with; the code ships but you want the note on record. |
-| APPROVE         | B- or above     | No significant concerns; the change is clean and correct. |
-| UNKNOWN         | —               | The diff was too truncated, context-free, or otherwise insufficient to assess. |
-
-**Keep your verdict consistent with your grade.** A grade of "D" must have verdict REQUEST_CHANGES;
-a grade of "F" must have verdict BLOCK; a grade of "B-" or above must have verdict APPROVE.
-
-- Your default verdict is APPROVE (default grade A-). You bear the burden of proof to escalate.
-- APPROVE* requires at least one Medium finding. Do not emit APPROVE* with only Low findings.
-- REQUEST_CHANGES requires ALL THREE: (a) a specific wrong line cited verbatim,
-  (b) a traceable failure path, (c) a concrete fix proposed.
-- Do NOT emit UNKNOWN just because the PR is large; use it only when you
-  genuinely cannot tell if the change is correct.
-- **Do not under-rate a clearly blocking issue as advisory.** If it would break
-  a build or corrupt data in production, assign severity=critical and verdict=BLOCK.
-
-## Compile-break rule (CRITICAL)
-If the diff REMOVES a symbol (enum value, method, constant, field, function
-signature change) AND the same diff still shows remaining references or
-call-sites to that removed symbol elsewhere in the codebase, that is a
-compile-time regression.  Assign the finding severity=critical and
-verdict=BLOCK (grade=F).  No other context softens this.
-
-## Severity anchors for findings
-Every finding MUST have a `severity` from:
-- **critical** — compile error, data corruption, security bypass, auth failure.
-- **high**     — confirmed correctness bug, silent data loss, unhandled exception
-  path, missing required migration, resource leak with real consequence.
-- **medium**   — advisory: code smell, suboptimal pattern, minor risk, the author
-  may reasonably disagree.
-- **low**      — cosmetic, documentation gap, style preference.
-
-## What to review
-Focus on: correctness bugs, security issues, data-loss risks, logic errors.
-Note but do not block on: style, minor naming, documentation gaps, test coverage.
-
-## Output (REQUIRED — populate the structured response fields)
-- `grade`: one of A+, A, A-, B+, B, B-, C+, C, C-, D+, D, D-, F.
-- `grade_justification`: one-sentence reason for the grade.
-- `verdict`: one of APPROVE, APPROVE*, REQUEST_CHANGES, BLOCK, UNKNOWN.
-- `summary`: one sentence summary of the review.
-- `findings`: array of issues found (empty array if none).
-  Each finding has: title, body (detailed description), severity (low/medium/high/critical),
-  confidence (0.0–1.0), file (source file path), line (null if not applicable).
-
-`confidence` is a float in [0.0, 1.0].
-`line` may be null if no specific line is applicable.
-`findings` may be an empty array if there are no issues."#
+/// Build the base system prompt with optional coverage-gating language.
+///
+/// Why: when coverage gating is enabled, the "do not block on coverage" advisory
+/// in the stock prompt becomes inaccurate (the runner WILL lower the verdict if
+/// coverage is insufficient).  This function is the single source of truth for
+/// both variants.
+/// What: when `coverage_gating_enabled` is false, the prompt is identical to the
+/// pre-#1014 stock text.  When true, the "Note but do not block on" coverage line
+/// is replaced with an informational note about the coverage context block.
+/// Test: `system_prompt_coverage_gating_on`, `system_prompt_coverage_gating_off`.
+pub fn reviewer_system_prompt_with_coverage(coverage_gating_enabled: bool) -> &'static str {
+    if coverage_gating_enabled {
+        SYSTEM_PROMPT_COVERAGE_GATING
+    } else {
+        SYSTEM_PROMPT_STOCK
+    }
 }
 
 // ─── Layered system prompt ────────────────────────────────────────────────────
@@ -228,10 +196,26 @@ Note but do not block on: style, minor naming, documentation gaps, test coverage
 /// What: appends principles then voice addenda to the stock base when they are
 /// non-empty; a blank separator line is inserted between layers.  When
 /// `voice_config` is all-None (stock-only), the output equals `reviewer_system_prompt()`.
+/// `coverage_gating_enabled` selects the stock base variant (#1014): when true the
+/// "do not block on coverage" advisory is replaced with an informational note.
 /// Test: `build_system_prompt_stock_only`, `build_system_prompt_with_principles`,
 /// `build_system_prompt_full_pipeline` in `prompt_tests.rs`.
 pub fn build_system_prompt(voice_config: &VoiceConfig) -> String {
-    let stock = reviewer_system_prompt();
+    build_system_prompt_with_coverage(voice_config, false)
+}
+
+/// Build the layered system prompt with an explicit coverage-gating flag.
+///
+/// Why: the runner calls this with `coverage_gating_enabled = config.coverage.enabled`
+/// so the system prompt accurately reflects whether coverage can gate the verdict.
+/// What: selects the stock base via `reviewer_system_prompt_with_coverage`, then
+/// appends the principles and voice addenda exactly as `build_system_prompt` does.
+/// Test: `build_system_prompt_coverage_gating_on`.
+pub fn build_system_prompt_with_coverage(
+    voice_config: &VoiceConfig,
+    coverage_gating_enabled: bool,
+) -> String {
+    let stock = reviewer_system_prompt_with_coverage(coverage_gating_enabled);
     let addendum = voice_config.combined_addendum();
     if addendum.is_empty() {
         return stock.to_string();
@@ -252,13 +236,16 @@ pub fn build_system_prompt(voice_config: &VoiceConfig) -> String {
 /// Bedrock tool-use or OpenRouter json_schema.
 /// `reviewer_model` may carry a `bedrock/` or `openrouter/` routing prefix;
 /// this function strips it before setting `LlmRequest.model`.
+/// `coverage_gating_enabled` selects the coverage-aware system prompt variant
+/// (#1014): when true, the "do not block on coverage" advisory is replaced.
 /// Test: `build_review_prompt_includes_diff`, `prompt_includes_context_blocks`,
 /// `build_review_prompt_strips_bedrock_prefix`,
 /// `build_review_prompt_includes_response_schema`,
 /// `build_review_prompt_with_voice_config_principles`,
-/// `build_review_prompt_with_voice_config_full`.
-// Eight arguments are required to fully specify the review (PR identity, diff,
-// context, model, voice).  The parameter count is structural, not incidental;
+/// `build_review_prompt_with_voice_config_full`,
+/// `build_review_prompt_coverage_gating_injects_block`.
+// Nine arguments are required to fully specify the review (PR identity, diff,
+// context, model, voice, coverage flag).  The parameter count is structural;
 // splitting would make the API less ergonomic without improving cohesion.
 #[allow(clippy::too_many_arguments)]
 pub fn build_review_prompt(
@@ -271,10 +258,73 @@ pub fn build_review_prompt(
     reviewer_model: &str,
     voice_config: &VoiceConfig,
 ) -> LlmRequest {
+    build_review_prompt_inner(
+        owner,
+        repo,
+        pr_meta,
+        diff,
+        context,
+        external_context,
+        reviewer_model,
+        voice_config,
+        false,
+    )
+}
+
+/// Build the `LlmRequest` with coverage-gating flag exposed (used by the runner).
+///
+/// Why: the runner calls this variant when `config.coverage.enabled` is true so
+/// the system prompt reflects that coverage can gate the verdict (#1014).
+/// What: identical to `build_review_prompt` but passes `coverage_gating_enabled`
+/// through to `build_system_prompt_with_coverage`.
+/// Test: `build_review_prompt_coverage_gating_injects_block`.
+#[allow(clippy::too_many_arguments)]
+pub fn build_review_prompt_with_coverage(
+    owner: &str,
+    repo: &str,
+    pr_meta: &ReviewPrMeta,
+    diff: &str,
+    context: &ReviewContext,
+    external_context: &str,
+    reviewer_model: &str,
+    voice_config: &VoiceConfig,
+    coverage_gating_enabled: bool,
+) -> LlmRequest {
+    build_review_prompt_inner(
+        owner,
+        repo,
+        pr_meta,
+        diff,
+        context,
+        external_context,
+        reviewer_model,
+        voice_config,
+        coverage_gating_enabled,
+    )
+}
+
+/// Internal implementation shared by both `build_review_prompt` variants.
+///
+/// Why: avoids code duplication between the public API-stable function and the
+/// coverage-aware variant while keeping the public interface clean.
+/// What: assembles the full `LlmRequest` from all inputs.
+/// Test: covered transitively by all `build_review_prompt_*` tests.
+#[allow(clippy::too_many_arguments)]
+fn build_review_prompt_inner(
+    owner: &str,
+    repo: &str,
+    pr_meta: &ReviewPrMeta,
+    diff: &str,
+    context: &ReviewContext,
+    external_context: &str,
+    reviewer_model: &str,
+    voice_config: &VoiceConfig,
+    coverage_gating_enabled: bool,
+) -> LlmRequest {
     let user_message = build_user_message(owner, repo, pr_meta, diff, context, external_context);
     LlmRequest {
         model: strip_provider_prefix(reviewer_model).to_string(),
-        system: build_system_prompt(voice_config),
+        system: build_system_prompt_with_coverage(voice_config, coverage_gating_enabled),
         messages: vec![ChatMessage {
             role: "user".to_string(),
             content: user_message,
@@ -323,152 +373,6 @@ impl ReviewPrMeta {
             url: result.pr_url.clone(),
         }
     }
-}
-
-/// Build the user-turn message for the review prompt.
-///
-/// Why: the user message carries all the review input: PR identity, diff, and
-/// context blocks.  Splitting it from the system prompt makes each independently
-/// tweakable.
-/// What: formats PR metadata as a header, then the diff block, then optional
-/// context sections for code search and static analysis, then any external
-/// context (`## Related <source>` markdown already rendered by the context
-/// orchestrator — JIRA / Confluence / GitHub Issues; APEX in PR-B).
-/// Test: `prompt_includes_context_blocks`, `prompt_includes_external_context`.
-fn build_user_message(
-    owner: &str,
-    repo: &str,
-    pr_meta: &ReviewPrMeta,
-    diff: &str,
-    context: &ReviewContext,
-    external_context: &str,
-) -> String {
-    let mut msg = String::with_capacity(diff.len() + 2048);
-
-    // PR header.
-    msg.push_str(&format!("## PR: {owner}/{repo}"));
-    if !pr_meta.title.is_empty() {
-        msg.push_str(&format!(" — {}", pr_meta.title));
-    }
-    msg.push('\n');
-    if !pr_meta.author.is_empty() {
-        msg.push_str(&format!("Author: @{}\n", pr_meta.author));
-    }
-    if !pr_meta.url.is_empty() {
-        msg.push_str(&format!("URL: {}\n", pr_meta.url));
-    }
-    msg.push('\n');
-
-    // Diff block.
-    msg.push_str("## Unified diff\n\n");
-    msg.push_str("```diff\n");
-    msg.push_str(diff);
-    if !diff.ends_with('\n') {
-        msg.push('\n');
-    }
-    msg.push_str("```\n\n");
-
-    // Code search context block.
-    if !context.search_results.is_empty() {
-        msg.push_str("## Related code (from trusty-search)\n\n");
-        for (i, result) in context.search_results.iter().enumerate().take(10) {
-            msg.push_str(&format!("### Context {} — {}\n", i + 1, result.file));
-            if let Some(ref snippet) = result.snippet {
-                msg.push_str("```\n");
-                msg.push_str(snippet);
-                if !snippet.ends_with('\n') {
-                    msg.push('\n');
-                }
-                msg.push_str("```\n");
-            }
-            msg.push('\n');
-        }
-    }
-
-    // Static-analysis context block.
-    if !context.complexity_hotspots.is_empty() {
-        msg.push_str("## Complexity hotspots (from trusty-analyze)\n\n");
-        for h in context.complexity_hotspots.iter().take(5) {
-            let fn_part = h
-                .function_name
-                .as_deref()
-                .map(|f| format!(" `{f}`"))
-                .unwrap_or_default();
-            msg.push_str(&format!(
-                "- `{}`{fn_part}: cyclomatic={}, cognitive={}\n",
-                h.file, h.cyclomatic, h.cognitive
-            ));
-        }
-        msg.push('\n');
-    }
-
-    if !context.smells.is_empty() {
-        msg.push_str("## Code smells (from trusty-analyze)\n\n");
-        for s in context.smells.iter().take(10) {
-            let line_part = s.line.map(|l| format!(" (line {l})")).unwrap_or_default();
-            msg.push_str(&format!(
-                "- `{}` — {} [{}]{line_part}\n",
-                s.file, s.category, s.severity
-            ));
-        }
-        msg.push('\n');
-    }
-
-    // APEX product-spec context block (Phase 6 PR-B, REV-420).
-    // Each result is a snippet from the spec/docs corpus that semantically
-    // matches the PR content.  Cite format: [apex: `path:line` — "excerpt"].
-    if !context.apex_results.is_empty() {
-        msg.push_str("## Related APEX product specs\n\n");
-        // defensive: apex_results already capped in fetch_apex_context; guard against future refactors
-        for (i, apex) in context
-            .apex_results
-            .iter()
-            .enumerate()
-            .take(crate::config::constants::MAX_APEX_RESULTS)
-        {
-            let line_suffix = apex.start_line.map(|l| format!(":{l}")).unwrap_or_default();
-            msg.push_str(&format!(
-                "### APEX {} — `{}{}`\n",
-                i + 1,
-                apex.file,
-                line_suffix
-            ));
-            if !apex.snippet.is_empty() {
-                msg.push_str("```\n");
-                msg.push_str(&apex.snippet);
-                if !apex.snippet.ends_with('\n') {
-                    msg.push('\n');
-                }
-                msg.push_str("```\n");
-            }
-            msg.push('\n');
-        }
-        msg.push_str(
-            "When citing an APEX spec, use the format: \
-             [apex: `path/to/spec.md:15` — \"brief excerpt\"]\n\n",
-        );
-    }
-
-    // External context block (rendered `## Related <source>` markdown from the
-    // context orchestrator — JIRA / Confluence / GitHub Issues).
-    // It is appended verbatim because the orchestrator already owns the heading
-    // + bullet format, keeping this builder source-agnostic.
-    let external = external_context.trim();
-    if !external.is_empty() {
-        msg.push_str(external);
-        if !external.ends_with('\n') {
-            msg.push('\n');
-        }
-        msg.push('\n');
-    }
-
-    // Structured-output instruction (schema-enforced; no need to emit a fence).
-    msg.push_str(
-        "Please review the diff above and populate the structured response \
-         fields (verdict, summary, findings) as specified in the system prompt.\n",
-    );
-
-    msg
 }
 
 // ─── Unit tests ───────────────────────────────────────────────────────────────
